hardware timer functionality
PC8 and PC9 now blink according to the hardware timer. At startup, the
code sets the pertinent registers such that the two LEDs blink
automatically without any additional code neccessary.

Timer3 is set to compare output mode with a period of 500ms. When the
compare is triggered, the status of the respective LED is toggled. PC8
is set to trigger 250 ms before PC9 does.

The commented out code is used to configue a software interrupt vector
to handle the timer overflowing. This is still an improvement over the
previous timer implementation which required a busy-wait to constantly
check the timer overflow flag.