push button debouncing
The push button incrementing propeller speed has been debounced. Instead
of relying on the hardware to trigger on a rising/ falling edge, we use
the TIM3 overflow interrupt to poll the input once every 20ms.

The debounce uses a hysteresis effect. While the button is pressed, a
static counter in the the overflow interrupt counts up. When the button
is depressed, it counts down. The counter is bounded to lie within -4
and 4. The speed is incremented if and only if the button is both
pressed down, and the counter equals 0. This ensures that continued
pressing of the button doesnt cause multiple increments.

The GPIO interrupts on the push button have now been disabled.